Hanbin: add a numerical result
