+++ src/alg/CellularAutomaton.jl
