+++ src/alg/readin.jl
