+++ src/alg/sddipParallel.jl
